--- src/main/java/com/sparta/givemetuna/domain/column/dto/StageResponseDto.java
@@ -1,4 +1,0 @@
-package com.sparta.givemetuna.domain.column.dto;
-
-public class StageResponseDto {
-}
