+++ src/main/java/com/sparta/givemetuna/domain/stage/dto/UpdateStageRequestDto.java
@@ -1,0 +1,4 @@
+package com.sparta.givemetuna.domain.stage.dto;
+
+public class UpdateStageRequestDto {
+}
