+++ src/main/java/com/sparta/givemetuna/domain/stage/repository/StageRepository.java
@@ -1,4 +1,4 @@
-package com.sparta.givemetuna.domain.column.repository;
+package com.sparta.givemetuna.domain.stage.repository;
 
 import com.sparta.givemetuna.domain.column.entity.Stage;
 import org.springframework.data.jpa.repository.JpaRepository;
